// src/interfaces/interface.tsx
export interface City {
  id: '2711537';
  name: 'Gothenburg';
  country: 'SE';
  coord: {
    lon: 11.9668;
    lat: 57.7072;
  };
  weather: {
    summary: {
      title: 'Clear';
      description: 'clear sky';
      icon: '01d';
    };
    temperature: {
      actual: 295.18;
      feelsLike: 294.61;
      min: 293.21;
      max: 296.86;
    };
    wind: {
      speed: 4.12;
      deg: 350;
    };
    clouds: {
      all: 0;
      visibility: 10000;
      humidity: 45;
    };
    timestamp: 1626686782;
  };
}